feat(home, setting): change logout button to setting screen
closes #31
