+++ .eslintrc.js
@@ -4,4 +4,7 @@ module.exports = {
   },
   root: true,
   extends: '@react-native-community',
+  rules: {
+    'no-console': 'error',
+  },
 };
