+++ src/authentication/index.ts
@@ -1,1 +1,2 @@
+export {default as logout} from './logout';
 export {default as useAuthentication} from './use-authentication';
