+++ src/screens/home/home.test.tsx
@@ -1,8 +1,6 @@
-import {GoogleSignin} from '@react-native-google-signin/google-signin';
 import {renderApp} from '@test';
-import {fireEvent, waitFor} from '@testing-library/react-native';
+import {fireEvent} from '@testing-library/react-native';
 import BottomSheet from 'react-native-bottomsheet';
-import Toast from 'react-native-toast-message';
 import HomeScreen from './home';
 
 const mockedNavigate = jest.fn();
@@ -14,16 +12,6 @@ jest.mock('@react-navigation/native', () => {
     }),
   };
 });
-const mockedSignOut = jest.fn();
-jest.mock('@react-native-firebase/auth', () => ({
-  __esModule: true,
-  default: () => ({
-    signOut: mockedSignOut,
-    onAuthStateChanged: jest.fn((callback) => {
-      callback({uid: 'USER_ID'});
-    }),
-  }),
-}));
 
 describe('Home Screen', () => {
   beforeEach(() => {
@@ -113,37 +101,5 @@ describe('Home Screen', () => {
     fireEvent.press(getByA11yLabel('Filter'));
 
     await expect(getByText('What I did Yesterday')).toBeDefined();
-  });
-
-  it('Logout Successfully', async () => {
-    jest.spyOn(GoogleSignin, 'revokeAccess').mockResolvedValue(null);
-    jest.spyOn(GoogleSignin, 'signOut').mockResolvedValue(null);
-
-    const {getByA11yLabel} = renderApp({
-      Component: HomeScreen.Component,
-      navigationOptions: HomeScreen.options,
-    });
-
-    fireEvent.press(getByA11yLabel('Logout'));
-
-    await waitFor(() => expect(mockedSignOut).toBeCalledTimes(1));
-  });
-
-  it('Logout Failed', async () => {
-    jest.spyOn(GoogleSignin, 'revokeAccess').mockResolvedValue(null);
-    jest.spyOn(GoogleSignin, 'signOut').mockResolvedValue(null);
-    mockedSignOut.mockRejectedValue(new Error('ERROR'));
-    const spyToastShow = jest.spyOn(Toast, 'show');
-
-    const {getByA11yLabel} = renderApp({
-      Component: HomeScreen.Component,
-      navigationOptions: HomeScreen.options,
-    });
-
-    fireEvent.press(getByA11yLabel('Logout'));
-
-    await waitFor(() => expect(mockedSignOut).toBeCalledTimes(1));
-    expect(spyToastShow).toBeCalledTimes(1);
-    expect(spyToastShow).toBeCalledWith({text2: 'ERROR', type: 'error'});
   });
 });
