+++ src/screens/home/home.tsx
@@ -1,7 +1,5 @@
 import {FAB, Header} from '@components';
 import styled from '@emotion/native';
-import auth from '@react-native-firebase/auth';
-import {GoogleSignin} from '@react-native-google-signin/google-signin';
 import {useNavigation} from '@react-navigation/native';
 import {useAppSelector} from '@store';
 import {Task, TaskSelectors, useFetchTask} from '@task';
@@ -11,7 +9,6 @@ import React, {useState} from 'react';
 import {FlatList} from 'react-native';
 import BottomSheet from 'react-native-bottomsheet';
 import {Appbar, useTheme} from 'react-native-paper';
-import Toast from 'react-native-toast-message';
 import {Task as TaskComponent} from './components';
 
 type Filter = keyof typeof FilterOptions;
@@ -77,19 +74,6 @@ const Component = (): JSX.Element => {
     );
   };
 
-  const onLogout = async () => {
-    try {
-      await GoogleSignin.revokeAccess().catch();
-      await GoogleSignin.signOut().catch();
-      await auth().signOut();
-    } catch ({message}) {
-      Toast.show({
-        type: 'error',
-        text2: message,
-      });
-    }
-  };
-
   return (
     <Screen>
       <Header>
@@ -98,11 +82,6 @@ const Component = (): JSX.Element => {
           accessibilityLabel="Filter"
           icon="filter-variant"
           onPress={onPresentFilterRelativeDay}
-        />
-        <Appbar.Action
-          accessibilityLabel="Logout"
-          icon="logout"
-          onPress={onLogout}
         />
       </Header>
       <FlatList
